Move add_vestigial_root to new arg_ops module
Create tsinfer/arg_ops.py for ARG topology operations and move
add_vestigial_root from matching.py. Create tests/test_arg_ops.py
with the corresponding tests. Update all call sites in matching.py,
test_python_c.py, and test_lshmm.py.
