+++ tests/test_matching.py
@@ -23,7 +23,6 @@ Tests for tsinfer.matching: make_root_ts, Matcher, extend_ts.
 from __future__ import annotations
 
 import numpy as np
-import pytest
 import tskit
 
 from tsinfer import grouping, matching, vcz
@@ -759,26 +758,6 @@ class TestMatcherExtendCycle:
         )
         assert "sequence_intervals" in ts2.metadata
         assert ts2.metadata["sequence_intervals"] == [[10, 51]]
-
-
-# ---------------------------------------------------------------------------
-# TestAddVestigialRoot
-# ---------------------------------------------------------------------------
-
-
-class TestAddVestigialRoot:
-    def test_non_discrete_genome(self):
-        tables = tskit.TableCollection(sequence_length=1.5)
-        tables.nodes.add_row(flags=tskit.NODE_IS_SAMPLE, time=0)
-        ts = tables.tree_sequence()
-        with pytest.raises(ValueError, match="discrete genome"):
-            matching.add_vestigial_root(ts)
-
-    def test_empty_tree_sequence(self):
-        tables = tskit.TableCollection(sequence_length=1)
-        ts = tables.tree_sequence()
-        with pytest.raises(ValueError, match="Emtpy trees"):
-            matching.add_vestigial_root(ts)
 
 
 # ---------------------------------------------------------------------------
